requirements.txt
asgiref==3.2.10
astroid==2.4.2
dj-database-url==0.4.2
Django==3.0.8
django-simple-bulma==2.0.3
djangorestframework==3.11.2
flake8==2.6.0
gunicorn==19.7.1
isort==5.6.4
lazy-object-proxy==1.4.3
libsass==0.20.1
mccabe==0.6.1
Pillow==7.2.0
pkg-resources==0.0.0
psycopg2==2.8.6
pycodestyle==2.0.0
pyflakes==1.2.3
pylint==2.6.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.2
whitenoise==5.2.0
wrapt==1.12.1
